*rewrite with JSX *refactor icon rotation logic
